+++ sdk/messaging/eventgrid/aznamespaces/models_serde.go
@@ -12,34 +12,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 )
 
-// MarshalJSON implements the json.Marshaller interface for type acknowledgeOptions.
-func (a acknowledgeOptions) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "lockTokens", a.LockTokens)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type acknowledgeOptions.
-func (a *acknowledgeOptions) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", a, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "lockTokens":
-			err = unpopulate(val, "LockTokens", &a.LockTokens)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", a, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type acknowledgeResult.
+// MarshalJSON implements the json.Marshaller interface for type AcknowledgeResult.
 func (a AcknowledgeResult) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "failedLockTokens", a.FailedLockTokens)
@@ -47,7 +20,7 @@ func (a AcknowledgeResult) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type acknowledgeResult.
+// UnmarshalJSON implements the json.Unmarshaller interface for type AcknowledgeResult.
 func (a *AcknowledgeResult) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
@@ -70,7 +43,7 @@ func (a *AcknowledgeResult) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type brokerProperties.
+// MarshalJSON implements the json.Marshaller interface for type BrokerProperties.
 func (b BrokerProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "deliveryCount", b.DeliveryCount)
@@ -78,7 +51,7 @@ func (b BrokerProperties) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type brokerProperties.
+// UnmarshalJSON implements the json.Unmarshaller interface for type BrokerProperties.
 func (b *BrokerProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
@@ -101,7 +74,7 @@ func (b *BrokerProperties) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type errorModel.
+// MarshalJSON implements the json.Marshaller interface for type Error.
 func (e Error) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "code", e.Code)
@@ -111,7 +84,7 @@ func (e Error) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type errorModel.
+// UnmarshalJSON implements the json.Unmarshaller interface for type Error.
 func (e *Error) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
@@ -134,7 +107,7 @@ func (e *Error) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type failedLockToken.
+// MarshalJSON implements the json.Marshaller interface for type FailedLockToken.
 func (f FailedLockToken) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "error", f.Error)
@@ -142,7 +115,7 @@ func (f FailedLockToken) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type failedLockToken.
+// UnmarshalJSON implements the json.Unmarshaller interface for type FailedLockToken.
 func (f *FailedLockToken) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
@@ -165,7 +138,7 @@ func (f *FailedLockToken) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type receiveDetails.
+// MarshalJSON implements the json.Marshaller interface for type ReceiveDetails.
 func (r ReceiveDetails) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "brokerProperties", r.BrokerProperties)
@@ -173,7 +146,7 @@ func (r ReceiveDetails) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type receiveDetails.
+// UnmarshalJSON implements the json.Unmarshaller interface for type ReceiveDetails.
 func (r *ReceiveDetails) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
@@ -196,14 +169,14 @@ func (r *ReceiveDetails) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type receiveResult.
+// MarshalJSON implements the json.Marshaller interface for type ReceiveResult.
 func (r ReceiveResult) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "value", r.Details)
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type receiveResult.
+// UnmarshalJSON implements the json.Unmarshaller interface for type ReceiveResult.
 func (r *ReceiveResult) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
@@ -223,34 +196,7 @@ func (r *ReceiveResult) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type rejectOptions.
-func (r rejectOptions) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "lockTokens", r.LockTokens)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type rejectOptions.
-func (r *rejectOptions) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", r, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "lockTokens":
-			err = unpopulate(val, "LockTokens", &r.LockTokens)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", r, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type rejectResult.
+// MarshalJSON implements the json.Marshaller interface for type RejectResult.
 func (r RejectResult) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "failedLockTokens", r.FailedLockTokens)
@@ -258,7 +204,7 @@ func (r RejectResult) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type rejectResult.
+// UnmarshalJSON implements the json.Unmarshaller interface for type RejectResult.
 func (r *RejectResult) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
@@ -281,34 +227,7 @@ func (r *RejectResult) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type releaseOptions.
-func (r releaseOptions) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "lockTokens", r.LockTokens)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type releaseOptions.
-func (r *releaseOptions) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", r, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "lockTokens":
-			err = unpopulate(val, "LockTokens", &r.LockTokens)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", r, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type releaseResult.
+// MarshalJSON implements the json.Marshaller interface for type ReleaseResult.
 func (r ReleaseResult) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "failedLockTokens", r.FailedLockTokens)
@@ -316,7 +235,7 @@ func (r ReleaseResult) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type releaseResult.
+// UnmarshalJSON implements the json.Unmarshaller interface for type ReleaseResult.
 func (r *ReleaseResult) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
@@ -339,34 +258,7 @@ func (r *ReleaseResult) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type renewLockOptions.
-func (r renewLockOptions) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "lockTokens", r.LockTokens)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type renewLockOptions.
-func (r *renewLockOptions) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", r, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "lockTokens":
-			err = unpopulate(val, "LockTokens", &r.LockTokens)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", r, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type renewLocksResult.
+// MarshalJSON implements the json.Marshaller interface for type RenewLocksResult.
 func (r RenewLocksResult) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "failedLockTokens", r.FailedLockTokens)
@@ -374,7 +266,7 @@ func (r RenewLocksResult) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type renewLocksResult.
+// UnmarshalJSON implements the json.Unmarshaller interface for type RenewLocksResult.
 func (r *RenewLocksResult) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {